requirements.txt
Django==4.1.3
djangorestframework==3.14.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
ijson==3.1.4